+++ redwood-layout-uiview/src/commonMain/kotlin/app/cash/redwood/layout/uiview/UIViewRedwoodLayoutWidgetFactory.kt
@@ -24,9 +24,11 @@ import app.cash.redwood.yoga.FlexDirection
 import platform.UIKit.UIView
 
 @ObjCName("UIViewRedwoodLayoutWidgetFactory", exact = true)
-public class UIViewRedwoodLayoutWidgetFactory : RedwoodLayoutWidgetFactory<UIView> {
+public class UIViewRedwoodLayoutWidgetFactory(
+  private val incremental: Boolean = false,
+) : RedwoodLayoutWidgetFactory<UIView> {
   override fun Box(): Box<UIView> = UIViewBox()
-  override fun Column(): Column<UIView> = UIViewFlexContainer(FlexDirection.Column)
-  override fun Row(): Row<UIView> = UIViewFlexContainer(FlexDirection.Row)
+  override fun Column(): Column<UIView> = UIViewFlexContainer(FlexDirection.Column, incremental)
+  override fun Row(): Row<UIView> = UIViewFlexContainer(FlexDirection.Row, incremental)
   override fun Spacer(): Spacer<UIView> = UIViewSpacer()
 }