+++ redwood-widget/src/iosMain/kotlin/app/cash/redwood/widget/UIViewChildren.kt
@@ -24,23 +24,24 @@ import platform.darwin.NSInteger
 @ObjCName("UIViewChildren", exact = true)
 public class UIViewChildren(
   private val container: UIView,
-  private val insert: (UIView, Modifier, Int) -> Unit = when (container) {
-    is UIStackView -> { view, _, index -> container.insertArrangedSubview(view, index.convert()) }
-    else -> { view, _, index -> container.insertSubview(view, index.convert<NSInteger>()) }
+  private val insert: (Widget<UIView>, UIView, Modifier, Int) -> Unit = when (container) {
+    is UIStackView -> { _, view, _, index -> container.insertArrangedSubview(view, index.convert()) }
+    else -> { _, view, _, index -> container.insertSubview(view, index.convert<NSInteger>()) }
   },
   private val remove: (index: Int, count: Int) -> Array<UIView> = when (container) {
     is UIStackView -> { index, count -> container.typedArrangedSubviews.remove(index, count) }
     else -> { index, count -> container.typedSubviews.remove(index, count) }
   },
   private val updateModifier: (Modifier, Int) -> Unit = { _, _ -> },
+  private val invalidateSize: () -> Unit = { (container.superview ?: container).setNeedsLayout() },
 ) : Widget.Children<UIView> {
   private val _widgets = ArrayList<Widget<UIView>>()
   override val widgets: List<Widget<UIView>> get() = _widgets
 
   override fun insert(index: Int, widget: Widget<UIView>) {
     _widgets.add(index, widget)
-    insert(widget.value, widget.modifier, index)
-    invalidate()
+    insert(widget, widget.value, widget.modifier, index)
+    invalidateSize()
   }
 
   override fun move(fromIndex: Int, toIndex: Int, count: Int) {
@@ -55,30 +56,38 @@ public class UIViewChildren(
     }
     subviews.forEachIndexed { offset, view ->
       val subviewIndex = newIndex + offset
-      insert(view, widgets[subviewIndex].modifier, subviewIndex)
+      val widget = widgets[subviewIndex]
+      insert(widget, view, widget.modifier, subviewIndex)
     }
-    invalidate()
+    invalidateSize()
   }
 
   override fun remove(index: Int, count: Int) {
+    for (i in index until index + count) {
+      val widget = _widgets[i]
+      if (widget is ResizableWidget<*>) {
+        widget.sizeListener = null // Break a reference cycle.
+      }
+    }
     _widgets.remove(index, count)
 
     remove.invoke(index, count)
-    invalidate()
+    invalidateSize()
   }
 
   override fun onModifierUpdated(index: Int, widget: Widget<UIView>) {
     updateModifier(widget.modifier, index)
-    invalidate()
-  }
-
-  private fun invalidate() {
-    (container.superview ?: container).setNeedsLayout()
+    invalidateSize()
   }
 
   override fun detach() {
     // Note that this doesn't update [container], since we don't want to trigger an update to the UI
     // if a detached widget is still on screen.
+    for (widget in _widgets) {
+      if (widget is ResizableWidget<*>) {
+        widget.sizeListener = null // Break a reference cycle.
+      }
+    }
     _widgets.clear()
   }
 }